Add debugging for split by seasons
The results differ a lot when splitting by seasons. I believe that the
data are the same (when using the clunky way of doing it by using
nearest index on dates and resampling on lon and lat), but splitting by
seasons cause it to grab different results. The original way use the
dates of sim_var and try to find the nearest index like that and split
seasons like that. My code uses the dates of obs_var (since we know the
dates) and split seasons like that.